+++ tests/unit-tests/index_test.cpp
@@ -54,7 +54,7 @@ TEST_F(IndexTest, Scan) {
     return false;
   });
   ASSERT_TRUE(count.has_value());
-  ASSERT_EQ(2, count.value());  // #Scan is not inclusive
+  ASSERT_EQ(size_t(2), count.value());  // #Scan is not inclusive
   db_->EndTransaction(tx, [](auto) {});
 }
 
@@ -65,7 +65,7 @@ TEST_F(IndexTest, AlphabeticalOrdering) {
 
   count = tx.Scan("carol", "zzz", [&](auto, auto) { return false; });
   ASSERT_TRUE(count.has_value());
-  ASSERT_EQ(1, count.value());
+  ASSERT_EQ(size_t(1), count.value());
   db_->EndTransaction(tx, [](auto) {});
 }
 
@@ -76,7 +76,7 @@ TEST_F(IndexTest, ScanViaTemplate) {
     return false;
   });
   ASSERT_TRUE(count.has_value());
-  ASSERT_EQ(2, count.value());
+  ASSERT_EQ(size_t(2), count.value());
   db_->EndTransaction(tx, [](auto) {});
 }
 
@@ -88,7 +88,7 @@ TEST_F(IndexTest, StopScanning) {
     return true;
   });
   ASSERT_TRUE(count.has_value());
-  ASSERT_EQ(1, count.value());
+  ASSERT_EQ(size_t(1), count.value());
   db_->EndTransaction(tx, [](auto) {});
 }
 
@@ -99,7 +99,7 @@ TEST_F(IndexTest, ScanWithoutEnd) {
     return false;
   });
   ASSERT_TRUE(count.has_value());
-  ASSERT_EQ(3, count.value());
+  ASSERT_EQ(size_t(3), count.value());
   db_->EndTransaction(tx, [](auto) {});
 }
